version 0.6

+++ src/drl_graph_3d.cpp
@@ -1,3 +1,35 @@
+/* 
+ * Copyright 2007 Sandia Corporation. Under the terms of Contract
+ * DE-AC04-94AL85000 with Sandia Corporation, the U.S. Government retains
+ * certain rights in this software.
+ * 
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without 
+ * modification, are permitted provided that the following conditions are 
+ * met:
+ * 
+ *     * Redistributions of source code must retain the above copyright 
+ * notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright 
+ * notice, this list of conditions and the following disclaimer in the 
+ * documentation and/or other materials provided with the distribution.
+ *     * Neither the name of Sandia National Laboratories nor the names of 
+ * its contributors may be used to endorse or promote products derived from 
+ * this software without specific prior written permission.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT 
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED 
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 // This file contains the member definitions of the master class
 
 #include <iostream>
@@ -11,7 +43,10 @@
 using namespace std;
 
 #include "drl_graph_3d.h"
-#include "random.h"
+#include "igraph_random.h"
+#include "igraph_interface.h"
+#include "igraph_progress.h"
+#include "igraph_interrupt_internal.h"
 #ifdef MUSE_MPI
   #include <mpi.h>
 #endif
@@ -745,7 +780,7 @@ void graph::Solve_Analytic( int node_ind, float &pos_x, float &pos_y,
    // Don't cut at end of scale
    if ( CUT_END >= 39500 ) return;
 
-   float num_connections = sqrtf(neighbors[node_ind].size());
+   float num_connections = (float)sqrt((float)neighbors[node_ind].size());
    float maxLength = 0;
 
    map<int, float>::iterator maxIndex;
